@@ -1,14 +1,18 @@
 {
   "name": "phanan/koel",
   "description": "Personal audio streaming service that works.",
-  "keywords": ["audio", "stream", "mp3"],
+  "keywords": [
+    "audio",
+    "stream",
+    "mp3"
+  ],
   "license": "MIT",
   "type": "project",
   "require": {
-    "php": ">=7.2.5",
-    "laravel/framework": "^7.0",
+    "php": ">=7.3",
+    "laravel/framework": "^8.0",
     "james-heinrich/getid3": "^1.9",
-    "guzzlehttp/guzzle": "^6.1",
+    "guzzlehttp/guzzle": "^7.0.1",
     "aws/aws-sdk-php-laravel": "^3.1",
     "pusher/pusher-php-server": "^4.0",
     "predis/predis": "~1.0",
@@ -21,16 +25,18 @@
     "laravel/helpers": "^1.0",
     "intervention/image": "^2.5",
     "laravel/sanctum": "^2.6",
-    "doctrine/dbal": "^2.10"
+    "doctrine/dbal": "^2.10",
+    "lstrojny/functional-php": "^1.14"
   },
   "require-dev": {
     "filp/whoops": "~2.0",
     "fzaninotto/faker": "~1.4",
     "mockery/mockery": "~1.0",
-    "phpunit/phpunit": "^8.5",
+    "phpunit/phpunit": "^9.0",
     "laravel/tinker": "^2.0",
     "php-mock/php-mock-mockery": "^1.3",
-    "friendsofphp/php-cs-fixer": "^2.16"
+    "friendsofphp/php-cs-fixer": "^2.16",
+    "dms/phpunit-arraysubset-asserts": "^0.2.1"
   },
   "suggest": {
     "ext-zip": "Allow downloading multiple songs as Zip archives"
@@ -41,7 +47,9 @@
     ],
     "psr-4": {
       "App\\": "app/",
-      "Tests\\": "tests/"
+      "Tests\\": "tests/",
+      "Database\\Factories\\": "database/factories/",
+      "Database\\Seeders\\": "database/seeders/"
     },
     "files": [
       "app/Helpers.php"
