@@ -5,21 +5,17 @@ namespace Tests\Feature;
 use App\Models\Song;
 use App\Models\User;
 use App\Services\LastfmService;
-use Exception;
 
 class ScrobbleTest extends TestCase
 {
-    /**
-     * @throws Exception
-     */
-    public function testLastfmScrobble()
+    public function testLastfmScrobble(): void
     {
         $this->withoutEvents();
         static::createSampleMediaSet();
 
         $song = Song::first();
         /** @var User $user */
-        $user = factory(User::class)->create();
+        $user = User::factory()->create();
         $user->setPreference('lastfm_session_key', 'foo');
 
         $timestamp = time();
