@@ -24,7 +24,7 @@ class DownloadAlbumCoverTest extends TestCase
 
     public function testHandle(): void
     {
-        $album = factory(Album::class)->make(['cover' => null]);
+        $album = Album::factory()->make(['cover' => null]);
         $event = new AlbumInformationFetched($album, ['image' => 'https://foo.bar/baz.jpg']);
 
         $this->mediaMetaDataService
