@@ -7,15 +7,11 @@ use App\Models\Interaction;
 use App\Models\Song;
 use App\Models\User;
 use App\Services\InteractionService;
-use Exception;
 use Illuminate\Support\Collection;
 use Tests\TestCase;
 
 class InteractionServiceTest extends TestCase
 {
-    /**
-     * @var InteractionService
-     */
     private $interactionService;
 
     public function setUp(): void
@@ -29,25 +25,19 @@ class InteractionServiceTest extends TestCase
     public function it_increases_a_songs_play_count(): void
     {
         /** @var Interaction $interaction */
-        $interaction = factory(Interaction::class)->create();
+        $interaction = Interaction::factory()->create();
 
         $this->interactionService->increasePlayCount($interaction->song, $interaction->user);
 
-        /** @var Interaction $interaction */
         $updatedInteraction = Interaction::find($interaction->id);
         self::assertEquals($interaction->play_count + 1, $updatedInteraction->play_count);
     }
 
-    /**
-     * @test
-     *
-     * @throws Exception
-     */
-    public function it_toggles_like_status(): void
+    public function testToggleLike(): void
     {
         $this->expectsEvents(SongLikeToggled::class);
 
-        $interaction = factory(Interaction::class)->create();
+        $interaction = Interaction::factory()->create();
 
         $this->interactionService->toggleLike($interaction->song, $interaction->user);
 
@@ -56,45 +46,38 @@ class InteractionServiceTest extends TestCase
         self::assertNotSame($interaction->liked, $updatedInteraction->liked);
     }
 
-    /**
-     * @test
-     *
-     * @throws Exception
-     */
-    public function user_can_like_multiple_songs_at_once(): void
+    public function testLikeMultipleSongs(): void
     {
         $this->expectsEvents(SongLikeToggled::class);
 
         /** @var Collection $songs */
-        $songs = factory(Song::class, 2)->create();
-        $user = factory(User::class)->create();
+        $songs = Song::factory(2)->create();
+
+        /** @var User $user */
+        $user = User::factory()->create();
 
         $this->interactionService->batchLike($songs->pluck('id')->all(), $user);
 
-        $songs->each(static function (Song $song) use ($user) {
+        $songs->each(static function (Song $song) use ($user): void {
             self::assertTrue(Interaction::whereSongIdAndUserId($song->id, $user->id)->first()->liked);
         });
     }
 
-    /**
-     * @test
-     *
-     * @throws Exception
-     */
-    public function user_can_unlike_multiple_songs_at_once(): void
+    public function testUnlikeMultipleSongs(): void
     {
         $this->expectsEvents(SongLikeToggled::class);
 
-        $user = factory(User::class)->create();
+        $user = User::factory()->create();
+
         /** @var Collection $interactions */
-        $interactions = factory(Interaction::class, 3)->create([
+        $interactions = Interaction::factory(3)->create([
             'user_id' => $user->id,
             'liked' => true,
         ]);
 
         $this->interactionService->batchUnlike($interactions->pluck('song.id')->all(), $user);
 
-        $interactions->each(static function (Interaction $interaction) {
+        $interactions->each(static function (Interaction $interaction): void {
             self::assertFalse(Interaction::find($interaction->id)->liked);
         });
     }
