@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 /**
@@ -13,10 +14,10 @@ use Illuminate\Database\Eloquent\Model;
  */
 class Setting extends Model
 {
+    use HasFactory;
+
     protected $primaryKey = 'key';
-
     public $timestamps = false;
-
     protected $guarded = [];
 
     /**
